Renames customerPage actions to footer
