+++ venia-integration-tests/cypress.config.json
@@ -15,7 +15,10 @@
     "scrollBehavior": "nearest",
     "trashAssetsBeforeRuns": false,
     "chromeWebSecurity": false,
-    "ignoreTestFiles": ["**/__snapshots__/*", "**/__image_snapshots__/*"],
+    "ignoreTestFiles": [
+        "**/__snapshots__/*", 
+        "**/__image_snapshots__/*"
+    ],
     "env": {
         "defaultData_couponCode": "PWA1423",
         "defaultData_giftCardNumber": "02Y0LCHUNDRZ",
