+++ venia-integration-tests/src/actions/footer/index.js
