+++ venia-integration-tests/src/actions/index.js
@@ -11,4 +11,4 @@ export * as pageBuilder from './pageBuilder';
 export * as productPage from './productPage';
 export * as toast from './toast';
 export * as wishlistPage from './wishlistPage';
-export * as customerPage from './customerPage';
+export * as footer from './footer';
